Prepare for 10.0.0-beta.4

+++ src/main/java/io/github/ennuil/ok_zoomer/zoom/overlays/ZoomerZoomOverlay.java
@@ -32,7 +32,7 @@ public class ZoomerZoomOverlay implements ZoomOverlay {
 		RenderSystem.depthMask(false);
 		RenderSystem.enableBlend();
 		RenderSystem.blendFuncSeparate(GlStateManager.SourceFactor.ZERO, GlStateManager.DestFactor.ONE_MINUS_SRC_COLOR, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
-		float fade = (float) transitionMode.getFade(deltaTracker.getGameTimeDeltaPartialTick(true));
+		float fade = transitionMode.getFade(deltaTracker.getGameTimeDeltaPartialTick(true));
 		RenderSystem.setShaderColor(fade, fade, fade, 1.0F);
 		graphics.blit(this.textureId, 0, 0, -90, 0.0F, 0.0F, graphics.guiWidth(), graphics.guiHeight(), graphics.guiWidth(), graphics.guiHeight());
 		RenderSystem.depthMask(true);
