+++ src/main/java/io/github/ennuil/ok_zoomer/zoom/transitions/SmoothTransitionMode.java
@@ -5,7 +5,7 @@ import net.minecraft.util.Mth;
 public class SmoothTransitionMode implements TransitionMode {
 	private boolean active;
 	private final float smoothMultiplier;
-	private double fovMultiplier;
+	private float fovMultiplier;
 	private float internalMultiplier;
 	private float lastInternalMultiplier;
 	private float internalFade;
@@ -30,13 +30,13 @@ public class SmoothTransitionMode implements TransitionMode {
 	}
 
 	@Override
-	public double applyZoom(double fov, float tickDelta) {
+	public float applyZoom(float fov, float tickDelta) {
 		fovMultiplier = Mth.lerp(tickDelta, this.lastInternalMultiplier, this.internalMultiplier);
 		return fov * fovMultiplier;
 	}
 
 	@Override
-	public double getFade(float tickDelta) {
+	public float getFade(float tickDelta) {
 		return Mth.lerp(tickDelta, this.lastInternalFade, this.internalFade);
 	}
